# doc/api/huilife-openapi.yml
openapi: 3.0.3
info:
  title: Huilife OpenAPI
  description: Huilife OpenAPI
  version: '1.0.20240415'
servers:
  - url: https://ingress.ai-messages.domain
...
paths:
...
  /shops/{shopId}/homeImages:
    get:
      summary: Find home images by shop ID
      operationId: getHomeImagesByShopId
      tags:
        - HomeImages
      description: Returns home images based on a single shop ID
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of home images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopHomeImages'
        '404':
          description: Home images not found
    post:
      summary: Add a new home image to a shop
      operationId: addHomeImageToShop
      tags:
        - HomeImages
      description: Adds a new home image to a shop based on a single shop ID
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: The created home image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopHomeImages'
    delete:
      summary: Delete all home images of a shop
      operationId: deleteAllHomeImagesOfShop
      tags:
        - HomeImages
      description: Deletes all home images of a shop based on a single shop ID
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Home images deleted
  /homeImages/{id}:
    get:
      summary: Find home image by ID
      operationId: getHomeImageById
      tags:
        - HomeImages
      description: Returns a home image object based on a single ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A home image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopHomeImages'
        '404':
          description: Home image not found
    put:
      summary: Update a home image
      operationId: updateHomeImage
      tags:
        - HomeImages
      description: Updates a home image based on a single ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopHomeImages'
      responses:
        '200':
          description: The updated home image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopHomeImages'
        '404':
          description: Home image not found
    delete:
      summary: Delete a home image
      operationId: deleteHomeImage
      tags:
        - HomeImages
      description: Deletes a home image based on a single ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Home image deleted

    ### introImages ###
  /shops/{shopId}/introImages:
    get:
      summary: Find intro images by shop ID
      operationId: getIntroImagesByShopId
      tags:
        - IntroImages
      description: Returns intro images based on a single shop ID
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of intro images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopIntroImages'
        '404':
          description: Intro images not found
    post:
      summary: Add a new intro image to a shop
      operationId: addIntroImageToShop
      tags:
        - IntroImages
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: The created intro image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopIntroImages'
    delete:
      summary: Delete all intro images of a shop
      operationId: deleteAllIntroImagesOfShop
      tags:
        - IntroImages
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Intro images deleted
  /introImages/{id}:
    get:
      summary: Find intro image by ID
      operationId: getIntroImageById
      tags:
        - IntroImages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: An intro image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopIntroImages'
        '404':
          description: Intro image not found
    put:
      summary: Update an intro image
      operationId: updateIntroImage
      tags:
        - IntroImages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopIntroImages'
      responses:
        '200':
          description: The updated intro image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopIntroImages'
        '404':
          description: Intro image not found
    delete:
      summary: Delete an intro image
      operationId: deleteIntroImage
      tags:
        - IntroImages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Intro image deleted

  ### shopRecommendations ###
  /shopRecommendations/{id}:
    get:
      summary: Find shop recommendation by ID
      operationId: getShopRecommendationById
      tags:
        - ShopRecommendations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A shop recommendation object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopRecommendations'
        '404':
          description: Shop recommendation not found
    put:
      summary: Update a shop recommendation
      operationId: updateShopRecommendation
      tags:
        - ShopRecommendations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopRecommendations'
      responses:
        '200':
          description: The updated shop recommendation object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopRecommendations'
        '404':
          description: Shop recommendation not found
    delete:
      summary: Delete a shop recommendation
      operationId: deleteShopRecommendation
      tags:
        - ShopRecommendations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Shop recommendation deleted
  /shopRecommendations:
    get:
      summary: Get all shop recommendations
      operationId: getAllShopRecommendations
      tags:
        - ShopRecommendations
      responses:
        '200':
          description: A list of shop recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShopRecommendations'
    post:
      summary: Create a new shop recommendation
      operationId: createShopRecommendation
      tags:
        - ShopRecommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopRecommendations'
      responses:
        '201':
          description: The created shop recommendation object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShopRecommendations'

  ### activities ###
...
  /images/{id}:
    get:
      summary: Find image by ID
      operationId: getImageById
      tags:
        - ActivityImages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: An image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityImages'
        '404':
          description: Image not found
    put:
      summary: Update an image
      operationId: updateImage
      tags:
        - ActivityImages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityImages'
      responses:
        '200':
          description: The updated image object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityImages'
        '404':
          description: Image not found
    delete:
      summary: Delete an image
      operationId: deleteImage
      tags:
        - ActivityImages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Image deleted

### components ###
components:
  schemas:
...
    ShopHomeImages:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        imageUrl:
          type: string
    ShopIntroImages:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        imageUrl:
          type: string
    ShopRecommendations:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        weight:
          type: integer
...
    ActivityImages:
      type: object
      properties:
        id:
          type: integer
        activityId:
          type: integer
        imageUrl:
          type: string